Add support for 'overrides' element to makefiles, along with '--overrides' option.
Includes are merged prior to the contents of the parent makefile, and as such,
the contents of the parent makefile take precedence. Order of defining projects
and includes in a makefile is irrelevent to this. This requires parent makefiles
as wrappers whenever we want to override an element from another makefile.

This commit introduces a new top-level element to makefiles: overrides. These
work identically to includes, except that they are merged over top of the
contents of the parent makefile.

This functionality becomes particularly useful when combined with the '--overrides'
option on the CLI. This allows, for example, environment-specific settings to
easily be passed in.